corrections ficier comments et mise en place du nécessaire pour les notifications

+++ controllers/comment.js
@@ -3,7 +3,7 @@ const xssFilters = require('xss-filters');
 
 
 exports.getAllComments = (req, res, next) => {
-    connexion.query(`SELECT * FROM comments WHERE postId = ?`, [req.body.publicationId], (error, result) => {
+    connexion.query(`SELECT id, postId, userName, date_comment, date_modif, content, modified, moderated FROM comments WHERE postId = ?`, [req.body.publicationId], (error, result) => {
         if(error) {
             res.status(500).send(error.code)
         }
@@ -11,8 +11,12 @@ exports.getAllComments = (req, res, next) => {
     })
   };
 
-exports.addComment = (req, res, next) => {
+  
 
+
+exports.addComment = (req, res, next) => {
+    const a = {...req.body};
+    console.log(a)
     connexion.query(`SELECT userId FROM users WHERE userName = ?`, [req.body.userName], (error, result) => {
         if(error) {res.status(500).send(error.sqlMessage)}
         else {
@@ -39,27 +43,50 @@ exports.addComment = (req, res, next) => {
 };
 
 exports.deleteComment = (req, res, next) => {
-    connexion.query(`DELETE FROM comments WHERE id=?`,[req.body.id], (error, result) => {
+
+    connexion.query(`SELECT userName FROM comments WHERE id = ?`, [req.body.id], (error, result) => {
         if(error) {res.status(500).send(error.sqlMessage)}
-        else {
-            connexion.query(`UPDATE publications 
-                SET numberComments = numberComments - 1
-                WHERE id = ?`, [req.body.postId], (error, result) => {
-                                                if (result) {res.status(200).send({message:"Comment deleted"});}
-                                                if (error) {res.status(500).send(error);}
-                                                })
-        }
-    })  
+        else if ( result.length !== 0 && result[0].userName === req.body.userName) {
+    
+          connexion.query(`DELETE FROM comments WHERE id=?`,[req.body.id], (error, result) => {
+                if(error) {res.status(500).send(error.sqlMessage)}
+                else {
+                    connexion.query(`UPDATE publications 
+                        SET numberComments = numberComments - 1
+                        WHERE id = ?`, [req.body.postId], (error, result) => {
+                                                        if (result) {res.status(200).send({message:"Comment deleted"});}
+                                                        if (error) {res.status(500).send(error);}
+                                                        })
+                }
+            })  
+    
+        } else {res.status(401).send({message:"Attention"})}
+      }
+      
+      )
+
+
+    
 };
 
 exports.modifyComment = (req, res, next) => {
-    const content = xssFilters.inHTMLData(req.body.content.replace(/\"/gi,'&µ'));
-    const modified = xssFilters.inHTMLData(req.body.modified);
-    const date_modif = xssFilters.inHTMLData(req.body.date_modif);
-    const id = xssFilters.inHTMLData(req.body.commentId);
-    connexion.query(`UPDATE comments SET content="${content}", modified="${modified}", date_modif="${date_modif}" WHERE id="${id}"`, (error, result) => {
+    
+    connexion.query(`SELECT userName FROM comments WHERE id = ?`, [req.body.commentId], (error, result) => {
         if(error) {res.status(500).send(error.sqlMessage)}
-        else {res.status(200).send({message:"Update done"})                                 
-        }
-    })  
+        else if (result.length !== 0 && result[0].userName === req.body.userName) {    
+            const content = xssFilters.inHTMLData(req.body.content.replace(/\"/gi,'&µ'));
+            const modified = xssFilters.inHTMLData(req.body.modified);
+            const date_modif = xssFilters.inHTMLData(req.body.date_modif);
+            const id = xssFilters.inHTMLData(req.body.commentId);
+            connexion.query(`UPDATE comments SET content="${content}", modified="${modified}", date_modif="${date_modif}" WHERE id="${id}"`, (error, result) => {
+                if(error) {res.status(500).send(error.sqlMessage)}
+                else {res.status(200).send({message:"Update done"})                                 
+                }
+            })    
+        } else {res.status(401).send({message:"Attention"})}
+      }
+      
+      )
+
+      
 };
